fix(F258): address sol review 2×P1 + 5×P2 findings
P1-1: fetchReleases settle pattern with response error/aborted + 30s timeout
P1-2: checkPendingUpgrade returns 'quitting' on retry-install, main.js
      guards against startAll() after recovery quit

P2-1: isReleaseComplete() — reject half-published releases missing
      platform assets (prevents offering partial uploads)
P2-2: skip-no-fallback — "Skip This Version" on highest version returns
      null instead of falling through to older versions
P2-3: checkDiskSpace() pre-download validation + extract getInstallType()
      and cleanUpdatesDir() to update-downloader.js (DRY + file-size)
P2-4: 6h periodic check timer (CHECK_INTERVAL_MS) for long-running sessions
P2-5: split update-checker.test.js (492→215+269 lines, both under 350)

96 tests pass, biome clean.

[宪宪/claude-opus-4-6🐾]
