+++ desktop/update-checker.test.js
@@ -1,83 +1,15 @@
-// F258 Phase A — update-checker unit tests (TDD)
+// F258 Phase A — update-checker unit tests: parsing, comparison, asset naming
+// Selection + persistence tests: see update-checker-selection.test.js
 const assert = require('node:assert/strict');
-const { mkdtempSync, rmSync, readFileSync, writeFileSync, mkdirSync } = require('node:fs');
-const { tmpdir } = require('node:os');
+const { readFileSync } = require('node:fs');
 const path = require('node:path');
-const { describe, test, beforeEach, afterEach } = require('node:test');
+const { describe, test } = require('node:test');
 
-const {
-  parseVersion,
-  compareSemver,
-  resolveAssetName,
-  extractAssetQuad,
-  selectUpdateTarget,
-  loadSettings,
-  saveSettings,
-  loadCachedReleases,
-  saveCachedReleases,
-} = require('./update-checker');
+const { parseVersion, compareSemver, resolveAssetName, extractAssetQuad } = require('./update-checker');
 
 function makeAsset(name, id, size, digest) {
   return { id, name, size, browser_download_url: `https://github.com/download/${name}`, digest };
 }
-
-const FIXTURE_RELEASES = [
-  {
-    tag_name: 'v0.12.0',
-    draft: false,
-    prerelease: false,
-    body: 'New features in 0.12.0',
-    assets: [
-      makeAsset('ClowderAI-Setup-0.12.0.exe', 201, 802000000, 'sha256:aaa111'),
-      makeAsset('ClowderAI-0.12.0-arm64.dmg', 202, 622000000, 'sha256:bbb222'),
-      makeAsset('ClowderAI-0.12.0-x64.dmg', 203, 632000000, 'sha256:ccc333'),
-    ],
-  },
-  {
-    tag_name: 'v0.13.0-beta.1',
-    draft: false,
-    prerelease: true, // ← should be filtered
-    body: 'Beta release',
-    assets: [
-      makeAsset('ClowderAI-Setup-0.13.0-beta.1.exe', 301, 810000000, 'sha256:ddd444'),
-      makeAsset('ClowderAI-0.13.0-beta.1-arm64.dmg', 302, 625000000, 'sha256:eee555'),
-      makeAsset('ClowderAI-0.13.0-beta.1-x64.dmg', 303, 635000000, 'sha256:fff666'),
-    ],
-  },
-  {
-    tag_name: 'v0.14.0',
-    draft: true, // ← should be filtered
-    prerelease: false,
-    body: 'Draft release',
-    assets: [
-      makeAsset('ClowderAI-Setup-0.14.0.exe', 401, 820000000, 'sha256:ggg777'),
-      makeAsset('ClowderAI-0.14.0-arm64.dmg', 402, 640000000, 'sha256:hhh888'),
-      makeAsset('ClowderAI-0.14.0-x64.dmg', 403, 650000000, 'sha256:iii999'),
-    ],
-  },
-  {
-    tag_name: 'v0.11.1',
-    draft: false,
-    prerelease: false,
-    body: 'Patch release',
-    assets: [
-      makeAsset('ClowderAI-Setup-0.11.1.exe', 101, 802000000, 'sha256:jjj000'),
-      makeAsset('ClowderAI-0.11.1-arm64.dmg', 102, 622000000, 'sha256:kkk111'),
-      makeAsset('ClowderAI-0.11.1-x64.dmg', 103, 632000000, 'sha256:lll222'),
-    ],
-  },
-  {
-    // Release with incomplete assets (missing Setup.exe) — should be skipped
-    tag_name: 'v0.11.2',
-    draft: false,
-    prerelease: false,
-    body: 'Incomplete release',
-    assets: [
-      makeAsset('ClowderAI-0.11.2-arm64.dmg', 501, 622000000, 'sha256:mmm333'),
-      // missing Setup.exe and x64.dmg
-    ],
-  },
-];
 
 describe('parseVersion', () => {
   test('parses v-prefixed tag', () => {
@@ -280,214 +212,4 @@ describe('extractAssetQuad', () => {
   });
 });
 
-describe('selectUpdateTarget', () => {
-  test('selects highest non-draft non-prerelease with complete assets (win)', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.10.1', 'win32', 'x64');
-    assert.notEqual(result, null);
-    assert.equal(result.version, '0.12.0');
-    assert.equal(result.asset.name, 'ClowderAI-Setup-0.12.0.exe');
-    assert.equal(result.asset.id, 201);
-    assert.equal(result.asset.digest, 'sha256:aaa111');
-    assert.equal(result.releaseNotes, 'New features in 0.12.0');
-    // P1-2: browser_download_url must survive through selectUpdateTarget
-    assert.equal(result.asset.browser_download_url, 'https://github.com/download/ClowderAI-Setup-0.12.0.exe');
-  });
-
-  test('selects highest for mac arm64', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.10.1', 'darwin', 'arm64');
-    assert.notEqual(result, null);
-    assert.equal(result.version, '0.12.0');
-    assert.equal(result.asset.name, 'ClowderAI-0.12.0-arm64.dmg');
-    assert.equal(result.asset.id, 202);
-  });
-
-  test('selects highest for mac x64', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.10.1', 'darwin', 'x64');
-    assert.notEqual(result, null);
-    assert.equal(result.version, '0.12.0');
-    assert.equal(result.asset.name, 'ClowderAI-0.12.0-x64.dmg');
-    assert.equal(result.asset.id, 203);
-  });
-
-  test('returns null when current version >= all releases', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.12.0', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('returns null when current version > all releases', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '1.0.0', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('filters out draft releases', () => {
-    // v0.14.0 is draft — should not be selected even though it's highest
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.12.0', 'win32', 'x64');
-    // 0.12.0 current, 0.14.0 is draft → no update available
-    assert.equal(result, null);
-  });
-
-  test('filters out prerelease', () => {
-    // v0.13.0-beta.1 is prerelease
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.12.0', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('skips releases with incomplete assets for the platform', () => {
-    // v0.11.2 missing Setup.exe → skipped for win32
-    // So win from 0.10.1 should get 0.12.0 (not 0.11.2)
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.10.1', 'win32', 'x64');
-    assert.equal(result.version, '0.12.0');
-  });
-
-  test('skips release when platform asset has no digest', () => {
-    const releasesNoDigest = [
-      {
-        tag_name: 'v0.12.0',
-        draft: false,
-        prerelease: false,
-        body: 'No digest',
-        assets: [
-          makeAsset('ClowderAI-Setup-0.12.0.exe', 201, 802000000, null),
-          makeAsset('ClowderAI-0.12.0-arm64.dmg', 202, 622000000, 'sha256:bbb222'),
-          makeAsset('ClowderAI-0.12.0-x64.dmg', 203, 632000000, 'sha256:ccc333'),
-        ],
-      },
-    ];
-    const result = selectUpdateTarget(releasesNoDigest, '0.10.1', 'win32', 'x64');
-    assert.equal(result, null, 'should skip release when target asset has no digest');
-  });
-
-  test('respects skippedVersion option', () => {
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.10.1', 'win32', 'x64', {
-      skippedVersion: '0.12.0',
-    });
-    // 0.12.0 is skipped → falls back to 0.11.1
-    assert.notEqual(result, null);
-    assert.equal(result.version, '0.11.1');
-  });
-
-  test('handles empty releases array', () => {
-    const result = selectUpdateTarget([], '0.10.1', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('handles releases with no assets', () => {
-    const emptyAssets = [{ tag_name: 'v0.12.0', draft: false, prerelease: false, body: '', assets: [] }];
-    const result = selectUpdateTarget(emptyAssets, '0.10.1', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('handles malformed tag_name gracefully', () => {
-    const badTags = [{ tag_name: 'not-a-version', draft: false, prerelease: false, body: '', assets: [] }];
-    const result = selectUpdateTarget(badTags, '0.10.1', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  // ── Pre-release currentVersion scenarios ──
-  test('finds update when current is pre-release of available release', () => {
-    // User has 0.12.0-rc.1 installed, v0.12.0 release exists → update available
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.12.0-rc.1', 'win32', 'x64');
-    assert.notEqual(result, null);
-    assert.equal(result.version, '0.12.0');
-  });
-
-  test('no update when current pre-release is higher than all releases', () => {
-    // User has 0.13.0-rc.1 installed, highest release is v0.12.0 → no update
-    const result = selectUpdateTarget(FIXTURE_RELEASES, '0.13.0-rc.1', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-
-  test('pre-release current does not downgrade to lower release', () => {
-    // User has 0.12.0-rc.1, v0.11.1 exists but is lower → should not offer
-    const onlyOlder = FIXTURE_RELEASES.filter((r) => r.tag_name === 'v0.11.1');
-    const result = selectUpdateTarget(onlyOlder, '0.12.0-rc.1', 'win32', 'x64');
-    assert.equal(result, null);
-  });
-});
-
-describe('settings persistence', () => {
-  let tempDir;
-
-  beforeEach(() => {
-    tempDir = mkdtempSync(path.join(tmpdir(), 'update-settings-'));
-  });
-
-  afterEach(() => {
-    rmSync(tempDir, { recursive: true, force: true });
-  });
-
-  test('loadSettings returns defaults when file does not exist', () => {
-    const settingsPath = path.join(tempDir, 'update-settings.json');
-    const settings = loadSettings(settingsPath);
-    assert.equal(settings.autoCheck, true);
-    assert.equal(settings.skippedVersion, null);
-    assert.equal(settings.lastCheckAt, null);
-    assert.equal(settings.etag, null);
-  });
-
-  test('saveSettings writes and loadSettings reads back', () => {
-    const settingsPath = path.join(tempDir, 'update-settings.json');
-    const data = {
-      autoCheck: false,
-      skippedVersion: '0.12.0',
-      lastCheckAt: '2026-07-07T00:00:00.000Z',
-      etag: '"abc123"',
-    };
-    saveSettings(settingsPath, data);
-    const loaded = loadSettings(settingsPath);
-    assert.deepEqual(loaded, data);
-  });
-
-  test('loadSettings tolerates corrupted JSON', () => {
-    const settingsPath = path.join(tempDir, 'update-settings.json');
-    writeFileSync(settingsPath, 'not valid json!!!');
-    const settings = loadSettings(settingsPath);
-    // Should return defaults, not throw
-    assert.equal(settings.autoCheck, true);
-    assert.equal(settings.skippedVersion, null);
-  });
-
-  test('saveSettings creates parent directory if needed', () => {
-    const settingsPath = path.join(tempDir, 'subdir', 'update-settings.json');
-    const data = { autoCheck: true, skippedVersion: null, lastCheckAt: null, etag: null };
-    saveSettings(settingsPath, data);
-    const content = JSON.parse(readFileSync(settingsPath, 'utf-8'));
-    assert.equal(content.autoCheck, true);
-  });
-});
-
-describe('release cache (P1-4: ETag/304 must not suppress Later re-prompt)', () => {
-  let tempDir;
-
-  beforeEach(() => {
-    tempDir = mkdtempSync(path.join(tmpdir(), 'release-cache-'));
-  });
-
-  afterEach(() => {
-    rmSync(tempDir, { recursive: true, force: true });
-  });
-
-  test('loadCachedReleases returns null when file missing', () => {
-    assert.equal(loadCachedReleases(path.join(tempDir, 'nope.json')), null);
-  });
-
-  test('saveCachedReleases + loadCachedReleases round-trip', () => {
-    const cachePath = path.join(tempDir, 'release-cache.json');
-    const releases = [{ tag_name: 'v0.12.0', draft: false, prerelease: false, body: 'test', assets: [] }];
-    saveCachedReleases(cachePath, releases);
-    const loaded = loadCachedReleases(cachePath);
-    assert.deepEqual(loaded, releases);
-  });
-
-  test('loadCachedReleases tolerates corrupted JSON', () => {
-    const cachePath = path.join(tempDir, 'release-cache.json');
-    writeFileSync(cachePath, '{{bad json');
-    assert.equal(loadCachedReleases(cachePath), null);
-  });
-
-  test('saveCachedReleases creates parent directory', () => {
-    const cachePath = path.join(tempDir, 'sub', 'release-cache.json');
-    saveCachedReleases(cachePath, []);
-    assert.deepEqual(loadCachedReleases(cachePath), []);
-  });
-});
+// Selection + persistence tests split into update-checker-selection.test.js
